requirements.txt
Flask==3.0.0
flask-cors==4.0.0
nltk==3.8.1
google-generativeai>=0.8.0